fix: set syscall exit arg as i32

+++ kernel/src/syscall/mod.rs
@@ -13,7 +13,7 @@ const SYSCALL_SBRK: usize = 214;
 pub fn syscall(id: usize, args: [usize; 3]) -> isize {
     match id {
         SYSCALL_WRITE => sys_write(args[0], args[1] as *const u8, args[2]),
-        SYSCALL_EXIT => sys_exit(args[0]),
+        SYSCALL_EXIT => sys_exit(args[0] as i32),
         SYSCALL_YIELD => sys_yield(),
         SYSCALL_GET_TIME => sys_get_time(args[0], args[1]),
         SYSCALL_SBRK => sys_sbrk(args[0] as i32),
