+++ kernel/src/syscall/process.rs
@@ -1,7 +1,7 @@
 use crate::{mm, task, timer};
 use log::{info, warn};
 
-pub fn sys_exit(exit_code: usize) -> ! {
+pub fn sys_exit(exit_code: i32) -> ! {
     match exit_code {
         0 => info!("Process exited with code {}", exit_code),
         _ => warn!("Process exited with code {}", exit_code),
